requirements.txt
streamlit>=1.50.0
streamlit-authenticator>=0.4.2
streamlit[pdf]
sqlalchemy
pydantic
pydantic[email]
cryptography
python-dotenv
aiosqlite
bcrypt
requests
pyyaml

# Documentation
mkdocs-material
mkdocs-with-pdf

# Development dependencies
pytest
pytest-asyncio
pytest-cov
black
isort
flake8
